requirements.txt
Flask==3.0.0
opencv-python-headless==4.11.0.86
paddleocr==2.10.0
paddlepaddle==2.6.2
ultralytics==8.3.90
numpy==1.26.4
gunicorn==23.0.0
arabic-reshaper==3.0.0
python-bidi==0.6.6